# Store vocabs in AnnifRegistry so they are shared between projects