Fixes #603

This PR makes the handling of vocabularies more efficient by making it possible to use a single AnnifVocabulary instance shared by multiple projects. The vocabularies are now stored in AnnifRegistry, similar to how projects are stored.

I did a little benchmarking. I set up TFIDF, MLLM and Parabel projects with YSO as the vocabulary, as well as an ensemble using these three as sources. I trained them using the archaeology test corpora from the test suite. Then I measured the user time and maximum RSS of two commands targeting the ensemble project: a simple `annif suggest` and a parallelized `annif eval -j4` command against the fulltext documents in the test suite, before and after this PR:

|           | time before | time after | RSS before | RSS after |
|-----------|-------------|------------|------------|-----------|
| suggest   | 6.81        | 5.51       | 438072     | 305828    |
| eval -j 4 | 24.35       | 23.24      | 418492     | 304464    |

This PR avoids loading the YSO vocabulary four times (once per project) and instead loads it only once. The result is 1.1-1.3 second reduction in CPU time and 110-130MB reduction in RAM usage.